Halfway done building payment command.
I have it accepting 2 inputs: user and payment amount.

with user it will search and return a list of potential matches; you have the option to choose up to 10 (max discord allows)
From there depending on how many you chose it will calc the total for each month (1,3,6,12) and do math to see what user paid for.
If no match it will take 1 month rate from paid amount
if there is extra (not a clean division) it takes that balance and evenly distrutes it to all users.

I need to build a discord view to display the info as currently its just a print which will allow proceed or cancel fork.
assuming proceed I then need to build the update db cmds
cancel will just end